requirements.txt
Flask==2.2.3
Flask_Login==0.6.2
flask_mysqldb==1.0.1
flask_sqlalchemy==3.0.3
h5py==3.8.0
Keras_Preprocessing==1.1.2
mysql_connector_repackaged==0.3.1
numpy==1.24.2
pandas==1.5.3
python-dotenv==0.21.1
requests==2.28.2
scikit_learn==1.2.1
tensorflow==2.11.0
Werkzeug==2.2.3
